Cover letter generator/requirements.txt
streamlit==1.18.1
python-dotenv==1.0.0
google-generativeai
altair==4.2.0
